refactor: notification 관련 컴포넌트들 폴더별로 관리하도록 수정
