+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/firebase.ts","./src/main.tsx","./src/vite-env.d.ts","./src/__test__/setup.ts","./src/apis/articleapi/article.ts","./src/apis/categoryapi/category.ts","./src/apis/searchssapi/recentsearch.ts","./src/apis/userapi/user.ts","./src/components/articleitem.tsx","./src/components/editquestiondialog.tsx.tsx","./src/components/errorboundary.tsx","./src/components/googlebutton.tsx","./src/components/header.tsx","./src/components/loading.tsx","./src/components/logintosignuptitle.tsx","./src/components/notificationbanner.tsx","./src/components/notificationwebapi.tsx","./src/components/reraclegametutorialmodal.tsx","./src/components/kakaoadfit/kakaoadfit320x100.tsx","./src/components/kakaoadfit/kakaoadfit320x50.tsx","./src/components/kakaoadfit/index.ts","./src/components/nav/nav.tsx","./src/components/nav/index.ts","./src/components/question/questionform.tsx","./src/components/question/questionitem.tsx","./src/components/question/index.ts","./src/components/wastecategory/category.tsx","./src/components/wastecategory/categorydetailitems.tsx","./src/components/wastecategory/categoryitems.tsx","./src/components/wastecategory/index.ts","./src/components/ui/alert.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/carousel.tsx","./src/components/ui/dialog.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/popover.tsx","./src/components/ui/select.tsx","./src/constant/reraclegameitem.ts","./src/constant/emailschema.ts","./src/constant/formatcontent.tsx","./src/constant/getfirsttextcontent.ts","./src/constant/loginschema.ts","./src/constant/signupschema.ts","./src/constant/__test__/getfirsttextcontent.test.ts","./src/constant/__test__/loginschema.test.ts","./src/constant/__test__/signupschema.test.ts","./src/layouts/detaillayout.tsx","./src/layouts/mainlayout.tsx","./src/layouts/navlayout.tsx","./src/layouts/index.ts","./src/lib/utils.ts","./src/lib/common/backheader/backheader.tsx","./src/lib/common/backheader/index.ts","./src/lib/common/categorybutton/announcementcategorybutton.tsx","./src/lib/common/categorybutton/postcategorybutton.tsx","./src/lib/common/categorybutton/index.ts","./src/lib/common/dropbox/dropbox.tsx","./src/lib/common/dropbox/index.ts","./src/lib/common/searchbar/searchbar.tsx","./src/lib/common/searchbar/index.ts","./src/pages/announcementdetailitempage.tsx","./src/pages/announcementpage.tsx","./src/pages/answerpage.tsx","./src/pages/articlepage.tsx","./src/pages/categorydetailitemspage.tsx","./src/pages/categoryitemspage.tsx","./src/pages/commentspage.tsx","./src/pages/homepage.tsx","./src/pages/loginpage.tsx","./src/pages/mypage.tsx","./src/pages/myquestionpage.tsx","./src/pages/notfoundpage.tsx","./src/pages/passwordresetpage.tsx","./src/pages/qnapage.tsx","./src/pages/reraclegame.tsx","./src/pages/signuppage.tsx","./src/pages/index.ts","./src/router/routes.tsx","./src/types/navbtntypes.ts","./src/types/reraclepuzzlegame.ts","./src/types/search.ts","./src/types/wastetypes.ts","./src/utils/articleservice.ts","./src/utils/chunkarray.ts","./src/utils/datekoreantime.ts","./src/utils/firestoreservice.ts","./src/utils/__test__/chunkarray.test.ts","./src/utils/__test__/datekoreantime.test.ts"],"errors":true,"version":"5.7.2"}
+{"root":["./src/app.tsx","./src/firebase.ts","./src/main.tsx","./src/vite-env.d.ts","./src/__test__/setup.ts","./src/apis/articleapi/article.ts","./src/apis/categoryapi/category.ts","./src/apis/searchssapi/recentsearch.ts","./src/apis/userapi/user.ts","./src/components/articleitem.tsx","./src/components/editquestiondialog.tsx.tsx","./src/components/errorboundary.tsx","./src/components/header.tsx","./src/components/loading.tsx","./src/components/reraclegametutorialmodal.tsx","./src/components/googlebutton/googlebutton.tsx","./src/components/kakaoadfit/kakaoadfit320x100.tsx","./src/components/kakaoadfit/kakaoadfit320x50.tsx","./src/components/kakaoadfit/index.ts","./src/components/nav/nav.tsx","./src/components/nav/index.ts","./src/components/notification/notificationbanner.tsx","./src/components/notification/notificationwebapi.tsx","./src/components/notification/index.ts","./src/components/question/questionform.tsx","./src/components/question/questionitem.tsx","./src/components/question/index.ts","./src/components/wastecategory/category.tsx","./src/components/wastecategory/categorydetailitems.tsx","./src/components/wastecategory/categoryitems.tsx","./src/components/wastecategory/index.ts","./src/components/ui/alert.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/carousel.tsx","./src/components/ui/dialog.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/popover.tsx","./src/components/ui/select.tsx","./src/constant/reraclegameitem.ts","./src/constant/emailschema.ts","./src/constant/formatcontent.tsx","./src/constant/getfirsttextcontent.ts","./src/constant/loginschema.ts","./src/constant/signupschema.ts","./src/constant/__test__/getfirsttextcontent.test.ts","./src/constant/__test__/loginschema.test.ts","./src/constant/__test__/signupschema.test.ts","./src/layouts/detaillayout.tsx","./src/layouts/mainlayout.tsx","./src/layouts/navlayout.tsx","./src/layouts/index.ts","./src/lib/utils.ts","./src/lib/common/backheader/backheader.tsx","./src/lib/common/backheader/index.ts","./src/lib/common/categorybutton/announcementcategorybutton.tsx","./src/lib/common/categorybutton/postcategorybutton.tsx","./src/lib/common/categorybutton/index.ts","./src/lib/common/dropbox/dropbox.tsx","./src/lib/common/dropbox/index.ts","./src/lib/common/searchbar/searchbar.tsx","./src/lib/common/searchbar/index.ts","./src/pages/announcementdetailitempage.tsx","./src/pages/announcementpage.tsx","./src/pages/answerpage.tsx","./src/pages/articlepage.tsx","./src/pages/categorydetailitemspage.tsx","./src/pages/categoryitemspage.tsx","./src/pages/commentspage.tsx","./src/pages/homepage.tsx","./src/pages/loginpage.tsx","./src/pages/mypage.tsx","./src/pages/myquestionpage.tsx","./src/pages/notfoundpage.tsx","./src/pages/passwordresetpage.tsx","./src/pages/qnapage.tsx","./src/pages/reraclegame.tsx","./src/pages/signuppage.tsx","./src/pages/index.ts","./src/router/routes.tsx","./src/types/category.ts","./src/types/navbtntypes.ts","./src/types/question.ts","./src/types/reraclepuzzlegame.ts","./src/types/search.ts","./src/types/wastetypes.ts","./src/utils/articleservice.ts","./src/utils/chunkarray.ts","./src/utils/datekoreantime.ts","./src/utils/firestoreservice.ts","./src/utils/__test__/chunkarray.test.ts","./src/utils/__test__/datekoreantime.test.ts"],"version":"5.7.2"}
